feat: add validation for open order

+++ src/cars/usecases/delete-car.usecase.ts
@@ -1,6 +1,7 @@
 import { AppError } from '@/common/domain/errors/app-error'
 import { CarModel } from '../domain/models/cars.model'
 import { CarsRepository } from '../domain/repositories/cars.repository'
+import { OrdersRepository } from '@/orders/domain/repositories/orders.repository'
 
 export type DeleteCarOutput = {
   id: string
@@ -16,7 +17,10 @@ export type DeleteCarOutput = {
 }
 
 export class DeleteCarUseCase {
-  constructor(private carRepository: CarsRepository) {}
+  constructor(
+    private carRepository: CarsRepository,
+    private orderRepository: OrdersRepository,
+  ) {}
 
   async execute(id: string): Promise<DeleteCarOutput> {
     const carExists = await this.carRepository.findById(id)
@@ -24,6 +28,11 @@ export class DeleteCarUseCase {
     if (!carExists) {
       throw new AppError("Car don't exist", 404)
     }
+
+    const order = await this.orderRepository.findWithCar(carExists.id)
+
+    if (order && order.status === 'Aberto')
+      throw new AppError('Car has a open order', 409)
     carExists.status = 'excluído'
 
     const deletedCar = await this.carRepository.delete(carExists)
